New version 16.2.0 (#13)
* "which" error no longers prints to console

* Silenced which command

* Spelling & imports fixed

* Another spelling fix

* Binary size reduction. From 8mb to 3mb

* Even smaller binary size. 2mb now

* readme remove dub todo check

* Change authors

* Separate utils into files & io (network requests)

* Spelling

* Grade_logic now also opens the IDE

* Added setup.sh to repo for easy install

* replace lazy_static to once_cell

* dep update

* Use semver for CLI version

* Added how to run instructions to readme.md and show full version with -V

* Even more instructions!

* Lms Folder command added to get the current assignment path

* Add install script

* Install add error handeling

* Install script executable command in readme

* Added module 'attempt'

* Removed Utils, wip self updating

* added upgrade script

* Allow debugging from subprocess

* remove old lms first

* Auto upgrade

* Added show path, & show settings. Refactored arguments a bit

* feat: update in lms, qol

* qol: change count to bool

* fix: install location before merge

* doc: settings options

* typo

* fix: default config true

* Spelling , logging update status and use of env!() for repository url

* About main command changed

* Refactored move_node_directories, thanks Stan!

* fix: remove autorelease

* feature: add functions and ascii to installer

* fix: error handle update

* fix: better error handeling

* Replace LICENSE

* Added webbrowser upload logic with settings toggle, refactored attempt.rs

new patch version because added functionality

* fix: verify version 16 compatible

* fix: v16 download all

* fix: readme 'upload_open_browser' option

* Version number and enable verify by lms open

* update: gitignore for "RustRover" editor

* fix: lms update remove current version when no internet

* fix: remove recursive retry update

* fix: install script

* Formatted from nightly + gitignore

* Spelling fixed

* python formatted + fixed error/warning messages

* Put Token in settings and made settings.config private. Fixed lms_ide bug

* Added new commands to toggle setups options

* update because of double pushes

* Update Readme + give options correct names

* Update check_todo to "setup" options so it is easy to toggle

* Strip arguments into multiple files so it is not that big anymore

* removed typo

* Show & Toggle

* Creating macros for easy (for me)  to use commands

* toggle shorter

* Toggle upgradable

* Update for loop

* fix: readme install endpoint

* fix: change git repo

* Update issue templates (#4) (#5)

* Update issue templates

* Update bug_report.md


* simple build github action

* open inner folder

* remove macros

* show commands

* Create github actions (#7)

* fix: workflow

* tmp

* tmp

* basic

* a

* a

* test2

* test2

* remove release

* a1

* rename

* it works i hope

* mac support

* sep job

* release build

* a

* a3

* a4

* outputs

* ftest1

* ftest2

* fix: set devops to main branch

---------


* toggle commands

* fix: install script

* test actions

* fix: install precompiled when update

* fix: action from nightly to main

* V16.1.0

* Cargo.lock version 16.1.0

* Formatter

* Removed unused env variable

* Format + make it visible what platform you have when not supported

* Check if path exe_name exist in path before trying to remove it

* Make the wget to a reqwest call

* remove: 32 bit processor support

* Setting the version number to the next version

* Updating documentation

* Updated README

* arch to Aarch

* Create test ini file for testing

* create function for couple settings that will be added for testing

* Change settings to mut reference so it is possible to use it for tests

* Allow instead of warn

* Add tests for files functions

* remove the time waiting (was because it had the same folder first)

* Test for installed commands

* One liner

* Smaller one liner

* Removed offline boolean, feature remains

* Ai review code (#9)

* feature: review code by ai

* fix: Cargo.toml error

* feature: ai review output

---------


* Macro's are back for checking if the command is equal to an existing command

* Made it more compact

* better format

* unwrap().to_string() -> ustring!()

* comments regarding the macros

* make oneliner

* ustr_ustring macro

* not tested for curl yet

* add mac_x64 support

* add curl command in readme

* test curl

* fix: add review to main.rs

* fix: show option to add lms to path

* fix: curl request in install script

* feature: support for linux arm

* fix: remove install script dependensies of wget and curl

* fix: review issues

---------

+++ src/io.rs
@@ -1,31 +1,31 @@
-use crate::files;
-
 use reqwest::{
     blocking::{Client, Response},
     StatusCode,
 };
 use serde_json::Value;
 use std::{
-    os::unix::fs::PermissionsExt,
-    env, fs,
+    env,
+    fs::{self, File},
     io::Write,
+    os::unix::fs::PermissionsExt,
     path::{Path, PathBuf},
     process::{exit, Command, Stdio},
 };
 
-// use crate::CLI_VERSION;
+const SUPPORTED_ARCHITECTURES: [&str; 2] = ["x86_64", "aarch64"];
 
+// TODO: Implement tests for request
 pub fn request(
     method: &str,
     path: String,
     token: &String,
     data: Option<Vec<u8>>,
 ) -> Option<Response> {
-    let url = if path.contains("?") {
-        format!("{}{}&v={}", crate::BASE_URL.to_string(), path, "999")
-    } else {
-        format!("{}{}?v={}", crate::BASE_URL.to_string(), path, "999")
-    };
+    let url = format!(
+        "{}{path}{}v=999",
+        crate::BASE_URL.to_string(),
+        if path.contains("?") { "&" } else { "?" }
+    );
 
     let client = Client::new();
 
@@ -74,6 +74,7 @@ pub fn request(
     }
 }
 
+// TODO: Implement tests for response_to_json
 pub fn response_to_json(res: Response) -> Value {
     let text = res.text().unwrap();
     match serde_json::from_str(&text) {
@@ -89,6 +90,9 @@ pub fn is_installed(application: &str) -> bool {
     return execute_command("which", vec![application]);
 }
 
+
+
+// TODO: Implement tests for execute_command
 pub fn execute_command(application: &str, args: Vec<&str>) -> bool {
     return Command::new(application)
         .args(args)
@@ -99,6 +103,7 @@ pub fn execute_command(application: &str, args: Vec<&str>) -> bool {
         .success();
 }
 
+// TODO: Implement tests for download_tgz
 pub fn download_tgz(path: String, token: &String, out_dir: &PathBuf) -> () {
     let res = request("GET", path, token, None);
 
@@ -141,6 +146,7 @@ pub fn download_tgz(path: String, token: &String, out_dir: &PathBuf) -> () {
     drop(tar_process)
 }
 
+// TODO: Implement tests for handle_upgrade
 pub fn handle_upgrade() {
     let exe_name = "lms";
 
@@ -153,29 +159,54 @@ pub fn handle_upgrade() {
 
     if !Path::exists(&lms_loc) {
         if let Err(err) = fs::create_dir_all(&lms_loc) {
-            eprintln!("A error occurred: {}", err);
+            eprintln!("A error occurred with creating: {}", err);
             exit(1)
         }
     }
 
-    if let Err(err) = fs::remove_file(&lms_loc.join(exe_name)) {
-        eprintln!("A error occurred: {}", err);
+    if Path::exists(&lms_loc.join(exe_name)) {
+        println!("Removing old version");
+        if let Err(err) = fs::remove_file(&lms_loc.join(exe_name)) {
+            eprintln!("A error occurred with removing: {}", err);
+            exit(1)
+        }
+    }
+
+    let architecture = std::env::consts::ARCH;
+
+    if !SUPPORTED_ARCHITECTURES.contains(&architecture) {
+        eprintln!("{} processor is not supported", &architecture);
         exit(1)
     }
 
     // TODO: Check if macos is arm or intel
-    let plat = match env::consts::OS { 
+    let plat = match env::consts::OS {
         "linux" => "linux_64",
         "macos" => "mac_arm64",
-        _ => {
-            eprintln!("Your platform is not supported");
+        platform => {
+            eprintln!("Your platform is not supported - {}", platform);
             exit(1)
         }
     };
-    
-    // TODO: Use reqwest instant of wget 
-    execute_command("wget", vec!["-q", "-O", lms_loc.join("lms").to_str().unwrap(), format!("https://github.com/gertjan84/lms-rust-cli/releases/latest/download/lms_{}", plat).as_str()]);
-    
-    fs::set_permissions(lms_loc.join("lms"), fs::Permissions::from_mode(0o755)).expect("Faild to set permissions");
+
+    let mut response = reqwest::blocking::get(format!(
+        "https://github.com/gertjan84/lms-rust-cli/releases/latest/download/lms_{}",
+        plat
+    ))
+    .expect("request failed");
+
+    if !response.status().is_success() {
+        eprintln!("Failed to download lms");
+        exit(1)
+    }
+
+    let mut file = File::create(lms_loc.join("lms")).expect("failed to create file");
+    response
+        .copy_to(&mut file)
+        .expect("failed to write to file");
+
+    fs::set_permissions(lms_loc.join("lms"), fs::Permissions::from_mode(0o755))
+        .expect("Failed to set permissions");
+
     println!("Installed");
 }
